Media app windows compatibility
