+++ public/apps/controller.html
@@ -35,7 +35,7 @@
   </div>
   <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.2/jquery.min.js"></script>
   <script src="/socket.io/socket.io.js"></script>
-  <script type="text/javascript" src="../js/controller.js"></script>
+  <script type="text/javascript" src="./controller.js"></script>
   <script type="text/javascript" src="../js/bottom.js"></script>
   <script type="text/javascript" src="../js/appcss.js"></script>
 </body>
