+++ public/apps/media.html
@@ -14,7 +14,7 @@
   </div>
   <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.2/jquery.min.js"></script>
   <script src="/socket.io/socket.io.js"></script>
-  <script type="text/javascript" src="../js/media.js"></script>
+  <script type="text/javascript" src="./media.js"></script>
   <script type="text/javascript" src="../js/bottom.js"></script>
   <script type="text/javascript" src="../js/appcss.js"></script>
 </body>
